36: Added support for literal under-scores in test names
Two or more consecutive under-scores will be printed in the output as a literal
under-score (rather than being converted to spaces as normal). Useful for
referring to functions in a non-confusing manner.

Example:

Test definition in test plan:
> function__name_returns_true() {

Test name printed in summary:
> function_name returns true

I may need to create like a ddt(5) manual to document this, i don't know.